refactor: rename pkg name to rankingservice
