+++ integration/rankingservice/route_by_osrm.go
@@ -1,4 +1,4 @@
-package ranking
+package rankingservice
 
 import (
 	"encoding/json"
